requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.1
flatbuffers==24.3.25
gast==0.5.4
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
joblib==1.4.2
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.14.1
